+++ js/modal.js
@@ -9,6 +9,7 @@
     onOpenEnd: null,
     onCloseStart: null,
     onCloseEnd: null,
+    preventScrolling: true,
     dismissible: true,
     startingTop: '4%',
     endingTop: '10%'
@@ -57,10 +58,7 @@
       this._openingTrigger = undefined;
       this.$overlay = $('<div class="modal-overlay"></div>');
 
-      Modal._increment++;
       Modal._count++;
-      this.$overlay[0].style.zIndex = 1000 + Modal._increment * 2;
-      this.el.style.zIndex = 1000 + Modal._increment * 2 + 1;
       this._setupEventHandlers();
     }
 
@@ -278,17 +276,24 @@
       }
 
       this.isOpen = true;
+      Modal._modalsOpen++;
+
+      // Set Z-Index based on number of currently open modals
+      this.$overlay[0].style.zIndex = 1000 + Modal._modalsOpen * 2;
+      this.el.style.zIndex = 1000 + Modal._modalsOpen * 2 + 1;
 
       // Set opening trigger, undefined indicates modal was opened by javascript
       this._openingTrigger = !!$trigger ? $trigger[0] : undefined;
-  
+
       // onOpenStart callback
       if (typeof(this.options.onOpenStart) === 'function') {
         this.options.onOpenStart.call(this, this.el, this._openingTrigger);
       }
 
-      let body = document.body;
-      body.style.overflow = 'hidden';
+      if (this.options.preventScrolling) {
+        document.body.style.overflow = 'hidden';
+      }
+
       this.el.classList.add('open');
       this.el.insertAdjacentElement('afterend', this.$overlay[0]);
 
@@ -312,6 +317,7 @@
       }
 
       this.isOpen = false;
+      Modal._modalsOpen--;
 
       // Call onCloseStart callback
       if (typeof(this.options.onCloseStart) === 'function') {
@@ -319,7 +325,11 @@
       }
 
       this.el.classList.remove('open');
-      document.body.style.overflow = '';
+
+      // Enable body scrolling only if there are no more modals open.
+      if (Modal._modalsOpen === 0) {
+        document.body.style.overflow = '';
+      }
 
       if (this.options.dismissible) {
         document.removeEventListener('keydown', this._handleKeydownBound);
@@ -336,7 +346,7 @@
    * @static
    * @memberof Modal
    */
-  Modal._increment = 0;
+  Modal._modalsOpen = 0;
 
   /**
    * @static