+++ js/select.js
@@ -11,28 +11,28 @@
    * @class
    *
    */
-  class Select extends Component {
+  class FormSelect extends Component {
     /**
-     * Construct Select instance
+     * Construct FormSelect instance
      * @constructor
      * @param {Element} el
      * @param {Object} options
      */
     constructor(el, options) {
-      super(Select, el, options);
+      super(FormSelect, el, options);
 
       // Don't init if browser default version
       if (this.$el.hasClass('browser-default')) {
         return;
       }
 
-      this.el.M_Select = this;
+      this.el.M_FormSelect = this;
 
       /**
        * Options for the select
-       * @member Select#options
+       * @member FormSelect#options
        */
-      this.options = $.extend({}, Select.defaults, options);
+      this.options = $.extend({}, FormSelect.defaults, options);
 
       this.isMultiple = this.$el.prop('multiple');
 
@@ -58,7 +58,7 @@
      */
     static getInstance(el) {
       let domElem = !!el.jquery ? el[0] : el;
-      return domElem.M_Select;
+      return domElem.M_FormSelect;
     }
 
     /**
@@ -67,7 +67,7 @@
     destroy() {
       this._removeEventHandlers();
       this._removeDropdown();
-      this.el.M_Select = undefined;
+      this.el.M_FormSelect = undefined;
     }
 
     /**
@@ -220,10 +220,12 @@
 
       // Initialize dropdown
       if (!this.el.disabled) {
+        let dropdownOptions = $.extend({}, this.options.dropdownOptions);
+
         if (this.isMultiple) {
-          this.options.dropdownOptions.closeOnClick = false;
+          dropdownOptions.closeOnClick = false;
         }
-        this.dropdown = M.Dropdown.init(this.input, this.options.dropdownOptions);
+        this.dropdown = M.Dropdown.init(this.input, dropdownOptions);
       }
 
       // Add initial selections
@@ -382,9 +384,9 @@
     }
   }
 
-  M.Select = Select;
+  M.FormSelect = FormSelect;
 
   if (M.jQueryLoaded) {
-    M.initializeJqueryWrapper(Select, 'select', 'M_Select');
+    M.initializeJqueryWrapper(FormSelect, 'formSelect', 'M_FormSelect');
   }
 }(cash));